requirements.txt
argcomplete==3.2.1
asgiref==3.7.2
boto3==1.34.15
botocore==1.34.15
certifi==2023.11.17
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.1
durationpy==0.6
hjson==3.1.0
idna==3.6
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.3
placebo==0.9.0
pynamodb==5.5.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.1
troposphere==4.5.3
typing_extensions==4.9.0
urllib3==1.26.18
Werkzeug==3.0.1
zappa==0.58.0
